test(126184): Corrigindo os testes para o método  generate_columns
