+++ sme_sigpae_api/medicao_inicial/__tests__/test_services/test_utils.py
@@ -1,6 +1,7 @@
 import pytest
 
 from sme_sigpae_api.medicao_inicial.services.utils import (
+    generate_columns,
     get_categorias_dietas,
     get_nome_periodo,
     update_dietas_alimentacoes,
@@ -299,3 +300,38 @@ def test_update_dietas_alimentacoes():
     assert isinstance(dietas_alimentacoes, dict)
     assert categoria_a_enteral_restricao in dietas_alimentacoes.keys()
     assert dietas_alimentacoes[categoria_a_enteral_restricao] == lista_alimentacoes
+
+
+def test_generate_columns(faixas_etarias_ativas):
+    faixas = [faixa.id for faixa in faixas_etarias_ativas]
+    dict_periodos_dietas = {
+        "Solicitações de Alimentação": ["kit_lanche", "lanche_emergencial"],
+        "INTEGRAL": faixas,
+        "Infantil MANHA": [
+            "lanche",
+            "lanche_4h",
+            "refeicao",
+            "total_refeicoes_pagamento",
+            "sobremesa",
+            "total_sobremesas_pagamento",
+        ],
+        "DIETA ESPECIAL - TIPO A": ["lanche", "lanche_4h", "refeicao"],
+        "DIETA ESPECIAL - TIPO B": ["lanche", "lanche_4h"],
+    }
+    colunas = generate_columns(dict_periodos_dietas)
+    assert isinstance(colunas, list)
+    assert len(colunas) == 21
+    assert sum(1 for tupla in colunas if tupla[0] == "INTEGRAL") == 8
+    assert sum(1 for tupla in colunas if tupla[0] == "Infantil MANHA") == 6
+    assert sum(1 for tupla in colunas if tupla[0] == "DIETA ESPECIAL - TIPO A") == 3
+    assert sum(1 for tupla in colunas if tupla[0] == "DIETA ESPECIAL - TIPO B") == 2
+    assert sum(1 for tupla in colunas if tupla[0] == "Solicitações de Alimentação") == 2
+
+    assert sum(1 for tupla in colunas if tupla[1] == "kit_lanche") == 1
+    assert sum(1 for tupla in colunas if tupla[1] == "lanche_emergencial") == 1
+    assert sum(1 for tupla in colunas if tupla[1] == "lanche") == 3
+    assert sum(1 for tupla in colunas if tupla[1] == "lanche_4h") == 3
+    assert sum(1 for tupla in colunas if tupla[1] == "refeicao") == 2
+    assert sum(1 for tupla in colunas if tupla[1] == "sobremesa") == 1
+    assert sum(1 for tupla in colunas if tupla[1] == "total_refeicoes_pagamento") == 1
+    assert sum(1 for tupla in colunas if tupla[1] == "total_sobremesas_pagamento") == 1
